# docs/swagger.yaml
basePath: /api/
definitions:
...
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
  models.Project:
    properties:
      creator_id:
        type: integer
      description:
        maxLength: 1024
        type: string
      id:
        type: integer
      name:
        maxLength: 64
        type: string
    type: object
...
  models.User:
    properties:
      address:
        maxLength: 100
        type: string
      admin:
        type: boolean
      email:
        maxLength: 60
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      patronymic:
        type: string
      surname:
        maxLength: 60
        minLength: 2
        type: string
    required:
    - address
    - name
    - password
    - surname
    type: object
  models.UserProductivity:
    properties:
      spent_hours:
        type: integer
      spent_minutes:
        type: integer
      task_id:
        type: integer
    type: object
...
  utils.Response:
    properties:
      ok:
        type: boolean
    type: object
...
host: localhost:80
info:
  contact:
    email: armangokka@gmail.com
    name: API Support
    url: https://t.me/armanokka
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time tracker REST API
  version: "1.0"
paths:
  /projects/:
    post:
      consumes:
      - application/json
      description: Create project
      parameters:
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Create project
      tags:
      - projects
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete project
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Delete project
      tags:
      - projects
    get:
      description: Get project by ID
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get project by ID
      tags:
      - projects
    patch:
      consumes:
      - application/json
      description: Update project info
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Project updates
        in: body
        name: project_updates
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Update project info
      tags:
      - projects
...
  /projects/{project_id}/users:
    get:
      description: Get project members list
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get project members list
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Add project member
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: User id of the member you want to invite
        in: body
        name: user_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Add project member
      tags:
      - projects
  /projects/{project_id}/users/{user_id}:
    delete:
      description: Remove project member
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: id of the user you want to remove
        in: path
        name: user_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Remove project member
      tags:
      - projects
    get:
      description: Get project member productivity
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: project id
        in: path
        name: user_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProductivity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get project member productivity
      tags:
      - projects
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
...